REF: move PythonImplementation into okonomiyaki.platforms.
