+++ okonomiyaki/file_formats/__init__.py
@@ -11,4 +11,3 @@ from ._egg_info import (
 )
 from ._package_info import PackageInfo
 from .egg import EggBuilder, EggRewriter
-from .pep425 import PythonImplementation
