+++ okonomiyaki/platforms/__init__.py
@@ -1,6 +1,7 @@
 # flake8: noqa
 from .epd_platform import EPDPlatform, applies
 from .platform import Platform
+from .python_implementation import PythonImplementation
 from ._arch import X86, X86_64
 
-__all__ = ["EPDPlatform", "Platform", "X86", "X86_64"]
+__all__ = ["EPDPlatform", "Platform", "PythonImplementation", "X86", "X86_64"]
